Migrate to NavigationSuiteScaffold
Currently missing animations
