+++ gradle/libs.versions.toml
@@ -53,6 +53,8 @@ composeCompiler = "1.5.14"
 composeFoundation = "1.7.0-beta05"
 composeMaterial = "1.7.0-beta05"
 composeMaterial3 = "1.3.0-beta04"
+composeMaterial3AdaptivenavigationSuite = "1.3.0-beta03"
+
 composeRuntime = "1.7.0-beta05"
 composeUi = "1.7.0-beta05"
 
@@ -100,6 +102,7 @@ compose-animation = { group = "androidx.compose.animation", name = "animation", 
 compose-foundation = { group = "androidx.compose.foundation", name = "foundation", version.ref = "composeFoundation" }
 compose-material = { group = "androidx.compose.material", name = "material", version.ref = "composeMaterial" }
 compose-material3 = { group = "androidx.compose.material3", name = "material3", version.ref = "composeMaterial3" }
+compose-material3-adaptiveNavigation = { group = "androidx.compose.material3", name = "material3-adaptive-navigation-suite", version.ref = "composeMaterial3AdaptivenavigationSuite" }
 compose-runtime = { group = "androidx.compose.runtime", name = "runtime", version.ref = "composeRuntime" }
 compose-ui = { group = "androidx.compose.ui", name = "ui", version.ref = "composeUi" }
 compose-ui-text-googleFonts = { group = "androidx.compose.ui", name = "ui-text-google-fonts", version.ref = "composeUi" }
